requirements.txt
pyotp
requests
pyyaml